Use modified version of mysql connector to work with sphinxQL
Using sphinx 2.0.4 and the 5.1.13 jdbc/mysql connector and not specifying
a mysql_version_string in the sphinx config we could connect and query, but
the results would not include the column names

Once we specified a mysql_version_string, we had connection problems (for
both sphinx 2.0.4 and 2.0.6).  Two problems were jdbc connector
requesting
select @@session.auto_increment_increment
and
show collations

Now, using sphinx 2.0.6 and setting a mysql_version_string in the sphinx
config (under searchd), eg
  mysql_version_string = 5.1.51
and this version of the connector then things should work (connections and
getting the column names in the results).

Not sure why this still doesn't work with 2.0.4 of sphinx.

NOTE: the 2.1 branch of sphinx (not yet released) supports enough to let
the jdbc driver connect and function.